+++ src/twg2/collections/dataStructures/PairList.java
@@ -11,7 +11,7 @@ import java.util.Map.Entry;
 import java.util.Set;
 
 import twg2.collections.interfaces.PairCollection;
-import twg2.collections.interfaces.PairCollectionImmutable;
+import twg2.collections.interfaces.PairCollectionReadOnly;
 import twg2.collections.util.ToStringUtil;
 
 /** Map implementation which allows duplicate keys and values 
@@ -333,7 +333,7 @@ public class PairList<K, V> implements PairCollection<K, V> {
 	 * @param listPairs pairList to add to this pairList
 	 */
 	@Override
-	public void putAll(PairCollectionImmutable<? extends K, ? extends V> listPairs) {
+	public void putAll(PairCollectionReadOnly<? extends K, ? extends V> listPairs) {
 		mod++;
 		keys.addAll(listPairs.keyList());
 		values.addAll(listPairs.values());